backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymupdf
python-multipart
pillow
requests
pydantic
crewai
crewai-tools
boto3
botocore
python-docx
pytesseract
tesseract
sqlalchemy
